requirements.txt
python-telegram-bot>=20.0
pytest>=9.0
pytest-asyncio>=0.22
python-dotenv>=1.0
